refactor(analysis): file the portable-name helpers where both archive writers can reach them
check:boundaries rejected the first placement. The helpers sat in analysis/zipArchive.ts, which is
analysis/ingest (tier 3), and caseArchive.ts is analysis/case (tier 1) — an import may go down a
tier or sideways, never up.

They are pure filename rules with no ZIP dependency, so they move down to the platform layer as
storage/portableFilename.ts. Both archive writers already import from storage/, and neither has to
depend on the other. Their tests move with them.

Also refreshes the cross-domain comply/total sentence in ARCHITECTURE.md, which
tests/architecture/moduleMap.test.ts asserts against the live scan: the new import adds one
complying edge, 1,911 of 1,950. The violation ledger is unchanged at 39.
